MSVC wants a cast-to-integer-type before dividing by a bool that should promote to int.  Sigh.  Followup to bug 851237, r=windows-redness
--HG--
extra : rebase_source : cf4ca438261d5bdf115a21d40f1ad4164ccbeaea

+++ js/public/TemplateLib.h
@@ -73,7 +73,7 @@ template <size_t N> struct NBitMask {
     // triggers divide-by-zero at compile time: a guaranteed compile error in
     // C++11, and usually one in C++98.  Add this value to |result| to assure
     // its computation.
-    static const size_t checkPrecondition = 0 / (N < BitSize<size_t>::result);
+    static const size_t checkPrecondition = 0 / size_t(N < BitSize<size_t>::result);
     static const size_t result = (size_t(1) << N) - 1 + checkPrecondition;
 };
 template <> struct NBitMask<BitSize<size_t>::result> {
